Move a VMS-specific include file in sv.c.
C++ doesn't allow pulling in rms.h smack dab in the middle of a
function like that.  While we're in there, go ahead and start
using the correct __VMS macro instead of just VMS.  The latter
was only necessary for VAX C, the non-C89-compliant predecessor
to DEC C that died a natural death some years ago.