Refactor some compiler-based hints in configure.com.
Once upon a time, there was VAX C, and there was a port of GCC
2.something that used the same library.  These both only ran on VAX
and both became obsolete years ago.  When DEC/Compaq/HP C came
along, it had its own library (typically referred to simply as
"the CRTL") that shipped with the operating system.  It was thus
natural to determine the presence or absence of a library function
based on the compiler choice, as that implied a library choice.

But there is now no choice of library nor has been for many years.
The CRTL is it, and even a new port of GCC (which apparently exists
in some form as part of Ada Core's toolchain) would almost
certainly use it.  So stop using the selected compiler to decide
whether a library function is present or not.