Put back an #if-0-ed chunk 7053d92 removed.
The chunk is not MAD-related but instead locale stuff.  I have no idea
why that chunk got removed (I used a combination of unifdef(1) and editor).
It's #if-0-ed, so no change of behavior either way, but let's keep
the code for now, since it seems to have "historical significance".