regcomp.c: Rewrite S_put_byte() which dumps regex []
This routine is used to display what the latin1 portions of a bracketed
character class matches.  It is what used, for example,  under -Dr to
display [] contents.  I noticed that the display had more hex characters
than needed, which are less intelligible than just printing out ASCII
printables.  And hence was harder to discern what is really going on.
This improves that, see the comments for more details.