Some more cppsymbols to scan.
Weirdly, BYTEORDER/BYTE_ORDER was missing.  Note that  __BYTE_ORDER__
might be '1234' even though longs are 64 bit, since probably the symbol
is used to tell the order in a "word", which still is "int", not "long".
In other words, Perl's BYTEORDER will likely be longer the the cppsymbol.

Gleaned some extra CPUs and OSes from
http://sourceforge.net/p/predef/wiki/Architectures/, though didn't go
all crazy on platforms with many variants like ARM, HPPA, SH.