POSIX::setlocale(LC_NUMERIC) returns proper locale
Commit 4c28b29c7a1549f173177cdf69e91cc1acaa0969 broke calling setlocale
with a single parameter and hoping to get the correct LC_NUMERIC
property returned from it (including LC_ALL calls).  This is because the
numeric locale is now kept as C except around the code that wants it to
be the underlying locale.  It turns out that querying the current locale
is such a code area, so LC_NUMERIC is now set to the underlying locale
around this.