t/run/locale.t: setting to POSIX may be same as C
On some (? all) systems, setting the locale to POSIX may cause it to
actually set it to C.  In these tests we want to exclude C, hence we
have to exclude POSIX as well.

Some systems may not have any other locales, so this commit also
properly skips tests on those.

Thanks to Matthew Horsfall for figuring this out