locale.c: Skip compiling fallback code on modern platforms
In the function that determines if a POSIX locale is UTF-8 or not, if
either nl_langinfo or MB_CUR_MAX are defined, it can reliably determine
the answer.  If they are not defined, it uses heuristics to figure
things out as best it can.  This code doesn't add value for those
platforms where one of the two symbols is defined, so can just be
ifdef'd out