locale.c: Refactor UTF8ness of currency symbol code
On older platforms that aren't C99 nor POSIX 2001, locale.c uses the
currency symbol to try to see if a locale is UTF-8 or not.  This commit
refactors it somewhat to make it cleaner, and which fixes several
problems.  The least issue was that it sometimes did a setlocale()
unnecessarily.  Others are that in some circumstances it called
localeconv() and/or looked at the result while within the wrong locale.