locale.c: Add new heuristic for finding if locale is UTF-8
On older platforms that don't conform to POSIX 2001 nor C99, heuristics
are employed to try to determine if a locale is UTF-8 or not.  This
commit improves those heuristics by looking at names of the months and
days of the week to see if they are UTF-8 or not.  This is done if
looking at the currency symbol failed to help.