locale.c: Fix some unused code for potential future use
This code extends the heuristics used to determine if a locale is UTF-8
or not on older platforms.  It has been #ifdef'd out because it only
added a little value on dromedary.  Now the previous commit has added
new heuristics, and tests on dromedary show that this adds nothing to
that.  But I'm leaving it in the source in case it might ever prove
useful.  In order to test it, I compiled it and found some problems with
the earlier version that this now fixes.