Fix legacy VMS feature checking status.
Back in 8dc9d3390b257b I consolidated two routines that were
almost identical -- except for their return values.  The routine
I kept returns the length of the equivalence name after logical
name translation, but some uses of it were checking it to see if
it was a successful VMS condition value.  Which means an odd length
(such as from "1") was successful but an even length (such as from
"ENABLE", the value recommended in the documentation) failed. So
fix those uses to check for a non-zero from simple_trnlnm.

For most features this only affected pre-7.3 systems, i.e., VMS
releases more than thirteen years old.  However, it also affected
features such as PERL_VMS_POSIX_EXIT that we have made up on our
own and are not tracking a CRTL feature.