Allow XSLoader to load modules from a different namespace
When trying to locate the binary component of a module, use the caller
package to find the library root directory, not the passed in module
name.  Since the file path it is using is always taken from the caller,
using the passed in package prevents it from finding the library root.
If in the package File::Spec::Unix, and given XSLoader::load("Cwd"), the
library root is three levels up from the filename of the
File::Spec::Unix module, not one.