Allow yylex to call the parser
yylex is expected to return exactly one token each time it is called.
Sometimes, it needs to remember a sequence of tokens, so these tokens
are stashed away in the parser struct and then emitted on the next
call (or next few calls) to yylex.

This makes yylex a lot more complicated than it would be if it could
call the parser and feed several tokens into it in a row.

This commit introduces a new parser function, yyemit, that is
intended to allow that.  As of yet, nothing calls it, so the mechan-
ism is untested.