Use YYEMIT for rightbracket
The code in toke.c under the ‘rightbracket’ label is more straight-
forward now.  The tokens are emitted in the order they appear in the
code, so you don’t have to tie your brain in knots to understand it.

I had to tweak the setting of PL_expect (aka PL_parser->expect) to
keep things working.

As it worked before, the parser (perly.y) would set PL_expect to
XSTATE after encountering a statement-terminating semicolon, which
would then be overridden by the the following '}' token (force_next
remembers the current value of PL_expect and it is set to that when
the next token is emitted).

If we are not using force_next, then YYEMIT(';') will make the parser
set PL_expect before we emit the '}', so the value yylex just set it
to will be clobbered.

It turns out that having the parser set it to XSTATE just makes things
more complicated.  We can instead have the lexer set it to XSTATE by
default when emitting an explicit semicolon.  (It was setting it to
XTERM.)  The parser can set it to XTERM in the only place that mat-
ters; viz., the header of a for-loop.

As a result, op.c:Perl_package and Perl_utilize no longer need to set
PL_expect at all.