toke.c: Have YYEMIT set PL_bufptr
This is necessary to keep the output of syntax errors the same
after the next commit.

PL_bufptr stores the current parsing position in the parser struct.
s is a variable local to yylex that stores it while parsing a token.
Usually the value is copied back to PL_bufptr when the function
returns, but we need to do it when calling yyemit as well to keep the
existing behaviour.