toke.c: Use YYEMIT for ‘foo @args’ sub calls
This fails ext/XS-APItest/t/swaptwostmts.t because of a little
blighter called yyunlex.

A keyword or function may have a custom parser that uses
parse_fullstmt and similar functions.  When parsing a statement such
as if(){...} without an else, the parser cannot know until it sees the
token after the final ‘}’ that it has reached the end of the state-
ment.  So parse_fullstmt puts that token back by calling yyunlex (this
call actually happens inside GRAMFULLSTMT, et al., in perly.y).

yyunlex uses the pending token stack (which I am trying to eliminate),
which causes problems for my new approach.

If the next statement after the if-without-else begins with something
the lexer emits as two tokens, then:

    • Previously, the lexer would put the second token on the pending
      token stack and return the first one.  Then yyunlex would take
      that first one and put it, too, on top of the pending token
      stack, so the next two yylex calls would pop that stack and Bob
      would be avuncular.

    • Now, the first token is emitted when yylex calls yyemit, which
      causes the parser to call yyunlex, putting that first token on
      the token stack.  When yyemit returns, then yylex returns the
      second token, which the parser handles while the first token is
      still pending.

That obviously can’t work, because the tokens come out backwards.

This ruins my plans to get rid of the pending token stack. :-(

So now I have to start from scratch....